+++ src/app/(root)/my-page/edit/page.tsx
@@ -1,4 +1,4 @@
-import MyEditPage from "@/components/page/my-edit-page";
+import MyEditPage from "@/pages/my-edit-page";
 
 const MyEditRoute = () => {
   return (
