+++ src/pages/gathering-detail-page.tsx
