+++ src/pages/my-page.tsx
